Make structure selectors use doubles for structure selection. Add overall spawn weight for structures.

+++ src/main/java/ivorius/reccomplex/structures/StructureInfo.java
@@ -29,7 +29,7 @@ public interface StructureInfo
 
     boolean isMirrorable();
 
-    int generationWeight(BiomeGenBase biome, WorldProvider provider);
+    double generationWeight(BiomeGenBase biome, WorldProvider provider);
 
     String generationCategory();
